refactor: move math functions to separate utilities class (#75)
